+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changes
 =======
 
+## UNRELEASED
+
+* Feature: Add `trace-deps` CLI.
+* Feature: Add `bailOnMissing` parameter to `traceFile`/`traceFiles`.
+* Feature: Add `dep` and `type` fields to `misses` array values returned by `traceFile`/`traceFiles`.
+
 ## 0.3.4
 
 * Bug: Handle non-`.js|mjs|json` extensions in JS files and parse when directly included. (E.g, `require('./url-alphabet/index.cjs')`).